add friendlyMaxVisitsFactor setting
Makes KataGo keep pace with the opponent by adjusting max visits
to how fast the opponent is playing.

Useful if you're running a ranked bot on a go server and want to be
friendly if the opponent is also playing fast, without sacrificing
playing strength if the opponent makes full use of the time control.

Using max visits makes the bot play naturally, responding quickly
if the move is expected and taking more time if it's unexpected.

(Of course, Katago's idea of what is obvious may differ from a human.)