Fix a Neural Engine overflow problem
Fix a Neural Engine overflow problem by a custom
Torch Mish activation that sets the threshold of
softplus to 11.

The threshold of softplus is modified to 11, which
is different from the original 20. This is because
exp(11) = 59874.14171519782 < 65504.0, so the
result of exp(11) can be represented by float16. If
the threshold of softplus is 20, the result of
exp(20) is 485165195.40979004, which is out of
range of float16.