Upgrade half library to 2.2
Compiling with
`cmake . -DUSE_BACKEND=CUDA -DCMAKE_CXX_FLAGS='-march=native'`
on a CPU with F16C intrinsics yielded the same errors as issue #387:

> error: ‘_mm_set_ss’ was not declared in this scope

The cause is that _mm_set_ss is defined in a header imported from
immintrin.h, which is included in
cpp/external/half-2.1.0/include/half.hpp:272.

However, it is only included if the preprocessor macro
HALF_ENABLE_F16C_INTRINSICS is set, but in turn, that is only set in
cpp/external/half-2.1.0/include/half.hpp:283, a few lines too late.

This issue seems fixed in half-2.2.0, thus we upgrade to it.