Improve model loading efficiency
Improve model loading efficiency by checking for existing compiled model and compiling if necessary.

This commit enhances the model loading process by introducing a functionality to verify the presence of a compiled model at a permanent URL. By doing so, I significantly improve the efficiency of loading models. In cases where a compiled model is not found, the function automatically attempts to compile the model using an ML package. If the compilation is successful, the function proceeds to create a permanent copy of the compiled model, which is subsequently utilized in future function calls.

To update the Core ML model, users are advised to remove any compiled ML models ("KataGoModel*.mlmodelc") located in the application support directory at `$HOME/Library/Application\ Support/`.