Scroll to last message and create message task
Scroll to the last message when the "messages" array changes, by using the ID of the last message. Also, create a message task on the initial view appearance to fetch messages from KataGo and continuously append them to the list of messages.

Created a infinite while loop in the "createMessageTask" function to continuously fetch and append new messages from KataGo.