Add more commands to ButtonView
This commit adds more commands to the `ButtonView` view. The commands are passed as an array and are used to create `CommandButton` instances. Each button executes its corresponding command when tapped. The `ButtonView` is now more flexible and can display a variable number of buttons based on the commands provided.

In addition, the commit also updates the `CommandView` to pass a new set of commands to the `ButtonView`.

The `GobanView` has been updated as well to include a new set of commands in the `ButtonView`.

The changes aim to enhance the functionality and usability of the UI by allowing users to easily execute a wider range of commands.