Refactor model name generation in CoreMLProcess
Previously, the model name was generated using a fixed string and some constants. Now, the model name is dynamically generated based on the maximum board length and precision information. The getModelName function has been added to achieve this. This commit improves the flexibility and accuracy of model name generation in CoreMLProcess.