Save and load NN outputs for cross-backend test
In order to improve testing and cross-backend checks, this change
introduces the ability to save and load base files for the GPU error
test. A new `baseFileName` argument is added to the `runFP16Test`
function in `testnnevalcanary.cpp`. When provided, this argument enables
the test to load a previously saved base file instead of recomputing the
base positions from scratch. This allows a backend to load a baseline NN
output file during testing.

The commit introduces two new functions in `testnnevalcanary.cpp`:
`saveBaseToFile` and `loadBaseFromFile`. The former takes a vector of
NNOutput and saves it to a binary file specified by `baseFileName`,
while the latter loads and populates a vector with NNOutput from a
binary file.

By allowing the test to load base files, it becomes easier to compare
and validate results between different backends. This enhances the
overall accuracy and reliability of the testing process.