Refactor getModelName to include length parameters
The getModelName method in the CoreMLBackend class has been modified to
accept additional xLen and yLen parameters. Previously, the method only
used a hardcoded value for board length. This change allows for
flexibility in specifying the board size when generating the model name.
The board length is now dynamically determined based on the xLen and
yLen parameters. This refactor improves code consistency and enhances
the reusability of the method.