Implement Torch Mish Operator with Sigmoid Approximation
This commit implements Torch Mish Operator with Sigmoid Approximation that can run on Neural Engine. In the previous revision, the built-in softplus exhibits great errors in 16-bit floating point computation, and the mish_torch_ne function that uses the `select` operation cannot run on Neural Engine anymore when upgrading macOS to the latest. This commit proposes a new softplus operation with sigmoid approximation for the mish activation function. The new softplus operation can run on Neural Engine and exhibit small error in 16-bit floating point computation efficiently and accurately.