Allow playing cleanup moves with rootPruneUselessMoves.
This allows analysis to rate/recommend these moves, even if one side is
passing. It also means cleanup will usually (but not *always*) proceed
through this path when playing via GTP, with the cleanup hack providing
backup only.

This shouldn't unduly affect the length of training games, since the
number of cleanup moves that can be played is strictly bounded, and most
training games are terminated as soon as all territory is pass-alive
anyway.