Speed up CoreML Model Loading
Updated the compileMLModel method to utilize a permanent URL for storing the compiled CoreML model. This change aims to improve loading times on subsequent launches by checking the permanent URL rather than recompiling the model each time.

This change is marked for caution as it was previously removed due to an unstable crash. Use this version with care.